# openapi.yaml
openapi: 3.1.0
info:
  title: Ledgerline
  version: 1.0.0
  description: Ledgerline is a recurring billing and payment collection API that lets small European businesses — gyms, clinics, schools, agencies — automatically charge their customers on a schedule without handling card data themselves. It sits between a merchant and a licensed payment provider, turning the painful, manual work of collecting recurring payments into something a business or their developer can plug in and trust.
  contact:
    name: Yassine Reguigui
    email: reguiguiyassine1@gmail.com
servers:
  - url: https://api.ledgerline.com/v1
    description: Production
    x-internal: false
  - url: https://sandbox.api.ledgerline.com/v1
    description: Sandbox
    x-internal: false
...
security:
  - apiKey: []
paths:
  /customers:
    get:
      tags:
        - customers
      summary: List Customers
      operationId: listCustomers
      description: "Returns a paginated list of customers in the current account, most recently created first. Page through results with `starting_after` set to the last id of the previous page. Read-only; no side effects."
      parameters:
        - $ref: "#/components/parameters/AfterParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags:
        - customers
      summary: Create Customer
      operationId: createCustomer
      description: "Creates a customer — the account holder that plans, subscriptions, and payment methods attach to. Requires an `Idempotency-Key` header so a retried request never creates a duplicate customer. Returns the full created customer, including its server-assigned `id`."
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerCreate"
      responses:
        "201":
          description: Created
          headers:
            Location:
              $ref: "#/components/headers/Location"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /customers/{customerId}:
    get:
      tags:
        - customers
      summary: Get a Customer
      operationId: getCustomer
      description: "Retrieves a single customer by id. Returns 404 if no customer with that id exists in the current account. Read-only."
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/CustomerId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: APIKey
      description: Authenticate with your API key. Use ll_live_... keys for production, ll_test_... for sandbox.
  schemas:
    Customer:
      type: object
      required:
        - id
        - object
        - created
        - livemode
        - email
      additionalProperties: true
      properties:
        id:
          $ref: "#/components/schemas/CustomerId"
        object:
          const: customer
          type: string
        created:
          type: string
          format: date-time
          maxLength: 40
        livemode:
          type: boolean
        email:
          type: string
          format: email
          maxLength: 255
        name:
          type: string
          maxLength: 255
          example: Will Smith
        description:
          type: ["string", "null"]
          maxLength: 500
        phone:
          type: ["string", "null"]
          maxLength: 32
        address:
          $ref: "#/components/schemas/Address"
        metadata:
          $ref: "#/components/schemas/Metadata"
    Address:
      type: object
      required:
        - country
      additionalProperties: false
      properties:
        country:
          type: string
          pattern: "^[A-Z]{2}$"
          maxLength: 2
          description: Two-letter ISO 3166-1 alpha-2 country code.
          example: NL
        line1:
          type: string
          maxLength: 255
        line2:
          type: ["string", "null"]
          maxLength: 255
        city:
          type: string
          maxLength: 255
        state:
          type: ["string", "null"]
          maxLength: 255
        postal_code:
          type: string
          maxLength: 16
    CustomerList:
      type: object
      required:
        - object
        - data
        - has_more
        - url
      properties:
        object:
          type: string
          const: "list"
        data:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/Customer"
        has_more:
          type: boolean
        url:
          type: string
          maxLength: 2048
          example: /v1/customers
    CustomerCreate:
      type: object
      additionalProperties: false
      required:
        - email
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        name:
          type: ["string", "null"]
          maxLength: 255
        description:
          type: ["string", "null"]
          maxLength: 500
        phone:
          type: ["string", "null"]
          maxLength: 32
        address:
          $ref: "#/components/schemas/Address"
        metadata:
          $ref: "#/components/schemas/Metadata"
    Metadata:
      type: object
      description: >-
        Key-value pairs you can attach to store your own identifiers or notes. Ledgerline never interprets these. Useful for linking a resource to a record in your system.
      maxProperties: 50
      propertyNames:
        maxLength: 40
      additionalProperties:
        type: string
        maxLength: 500
...
    CustomerId:
      type: string
      pattern: "^cus_[a-zA-Z0-9]+$"
      example: cus_abc123
      maxLength: 255
...
    ProblemDetail:
      type: object
      required:
        - type
        - title
        - status
        - code
      additionalProperties: true
      properties:
        code:
          type: string
          enum:
            - unauthorized
            - customer_not_found
            - validation_error
            - not_found
            - method_not_allowed
            - internal_error
          maxLength: 64
          description: |-
            Stable, machine-readable error code — the value clients branch on.

            | code | HTTP | type URI | when it occurs |
            | --- | --- | --- | --- |
            | `unauthorized` | 401 | `/problems/unauthorized` | No valid API key was provided. |
            | `customer_not_found` | 404 | `/problems/customer-not-found` | No customer exists with that id. |
            | `validation_error` | 422 | `/problems/validation-error` | Request body failed field validation. |
            | `not_found` | 404 | `about:blank` | The route does not exist. |
            | `method_not_allowed` | 405 | `about:blank` | The route exists but not for this method. |
            | `internal_error` | 500 | `about:blank` | An unexpected fault on our side. |

            The registry grows as handlers land.
          example: customer_not_found
        type:
          type: string
          format: uri
          maxLength: 512
          example: "https://api.ledgerline.com/problems/customer-not-found"
        title:
          type: string
          maxLength: 255
          example: "Customer Not Found"
        status:
          type: integer
          format: int32
          minimum: 100
          maximum: 599
          example: 404
        detail:
          type: string
          maxLength: 1000
          example: "No customer with id cus_abc123 exists in this account."
        instance:
          type: string
          format: uri-reference
          maxLength: 512
          example: "/v1/customers/cus_abc123"
    ValidationProblemDetail:
      allOf:
        - $ref: "#/components/schemas/ProblemDetail"
        - type: object
          properties:
            errors:
              type: array
              maxItems: 100
              items:
                type: object
                properties:
                  field:
                    type: string
                    maxLength: 255
                    example: email
                  code:
                    type: string
                    maxLength: 64
                    enum:
                      - required
                      - unknown_field
                      - invalid_type
                      - invalid
                    description: >-
                      Owned, stable validation code for this field, mapped from
                      the validation layer to a fixed vocabulary so it does not
                      shift when the framework changes.
                    example: required
                  message:
                    type: string
                    maxLength: 1000
                    example: Field required
                required:
                  - field
                  - code
                  - message
          required:
            - errors
  headers:
    Location:
      description: URL of the newly created resource
      required: true
      schema:
        type: string
        format: uri-reference
        maxLength: 2048
      example: /v1/customers/cus_abc123
  parameters:
    IdempotencyKey:
      description: "A unique string per request, max 255 chars. Resubmitting with the same key returns the cached response without re-executing the operation."
      in: header
      schema:
        type: string
        minLength: 1
        maxLength: 255
      name: Idempotency-Key
      required: true
    LimitParam:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        default: 20
    AfterParam:
      in: query
      name: starting_after
      required: false
      schema:
        type: string
        maxLength: 500
...
  responses:
    BadRequest:
      description: "The request was malformed — missing required headers, invalid JSON, or a field failed basic format validation. Check the detail field for specifics."
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    Unauthorized:
      description: "No valid API key was provided. Include your key in the Authorization: Bearer ll_live_... header."
      headers:
        WWW-Authenticate:
          description: The authentication scheme to use, as required by RFC 9110 for 401 responses.
          schema:
            type: string
            maxLength: 256
          example: Bearer
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    NotFound:
      description: "The requested resource does not exist or has been deleted."
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    UnprocessableEntity:
      description: "The request was well-formed but failed semantic validation — for example, a required field is present but its value is logically invalid. Fix the values in detail before retrying."
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ValidationProblemDetail"
    TooManyRequests:
      description: "You have exceeded the rate limit. Wait for the duration in the Retry-After header before retrying."
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
      headers:
        Retry-After:
          description: Seconds to wait before retrying.
          schema:
            type: integer
            format: int32
            minimum: 0
            maximum: 86400
        RateLimit-Limit:
          description: The request quota for the current window.
          schema:
            type: integer
            format: int32
            minimum: 0
            maximum: 1000000
        RateLimit-Remaining:
          description: Requests remaining in the current window.
          schema:
            type: integer
            format: int32
            minimum: 0
            maximum: 1000000
        RateLimit-Reset:
          description: Seconds until the quota resets.
          schema:
            type: integer
            format: int32
            minimum: 0
            maximum: 86400
    InternalServerError:
      description: "An unexpected error occurred on Ledgerline's side. Retrying with exponential backoff is safe. If the problem persists, contact support with the instance value from the response body."
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"